find out how much extended memory cycles can be sped up
the real PDP-8/L will run at approx 900nS

z8lpanel:
    hardreset               reset everything
    pin set xm_enlo4k 1     force all memory accesses to fpga ram
    fastxmem                speed up fpga ram cycles

+++ zynq/makebootbin.sh
@@ -19,8 +19,11 @@ cd ../zplin
 source petadevel/settings.sh
 source /tools/Xilinx/Vivado/2018.3/settings64.sh
 cd petakernl/zturn
-rm -f images/linux/BOOT.BIN
-petalinux-package --boot --force --fsbl ./images/linux/zynq_fsbl.elf --fpga $mydir/pdp8l.runs/impl_1/myboard_wrapper.bit --u-boot
+if [ images/linux/BOOT.BIN -ot $mydir/pdp8l.runs/impl_1/myboard_wrapper.bit ]
+then
+    rm -f images/linux/BOOT.BIN
+    petalinux-package --boot --force --fsbl ./images/linux/zynq_fsbl.elf --fpga $mydir/pdp8l.runs/impl_1/myboard_wrapper.bit --u-boot
+fi
 ls -l `pwd`/images/linux/BOOT.BIN
 scp images/linux/BOOT.BIN root@zturn$hostnum:/boot/BOOT.BIN
 set +e
